+++ lib/data/source/errors/api_error.dart
@@ -4,6 +4,9 @@ import 'package:flutter/foundation.dart';
 import 'package:get/get.dart';
 import 'package:http/http.dart' as http;
 import '../../../../utils/services/helpers.dart';
+import '../../../../utils/services/localstorage/hive.dart';
+import '../../../../utils/services/localstorage/keys.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import '../../../views/screens/auth/login_screen.dart';
 
 class ApiResponse {
@@ -21,7 +24,9 @@ class ApiResponse {
           return response;
 
         case 401:
-          Get.offAll(() => const LoginScreen());
+          if (FirebaseAuth.instance.currentUser == null && (HiveHelp.read(Keys.token) == null || HiveHelp.read(Keys.token).toString().isEmpty)) {
+            Get.offAll(() => const LoginScreen());
+          }
           return _logError(
             STATUS_CODE,
             URL,